CentraalMuseumUtrechtItem: changed original_object_urls
Centraal Museum Utrecht does host an individual page for each object in the
collection on their site. For each object, original_object_urls now
includes the URL to these HTML pages. The URL to the XML file was
removed from original_object_urls, since it pointed to a large XML file
that contained all records in the collection (instead of a single
object).